Refactored into two crates
