+++ Cargo.toml
@@ -1,21 +1,5 @@
-[package]
-name = "jasondb"
-version = "0.1.2"
-authors = ["William Henderson <william-henderson@outlook.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "3.0.0-beta.2", features = ["yaml"] }
-ctrlc = { version = "3.0", features = ["termination"] }
-dotenv = "0.15.0"
-futures = { version = "0.3.0", features = ["thread-pool"] }
-native-tls = "0.2.7"
-parking_lot = "0.10.2"
-serde = "1.0"
-serde_json = "1.0"
-tar = "0.4.33"
-tokio = { version = "1.5.0", features = ["full", "tracing"] }
-tokio-util = { version = "0.6.3", features = ["full"] }
-websocket = "0.21.1"
+[workspace]
+members = [
+  "jasondb",
+  "jasondb-server"
+]
