+++ jasondb-server/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "jasondb_server"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { version = "3.0.0-beta.2", features = ["yaml"] }
+ctrlc = { version = "3.0", features = ["termination"] }
+dotenv = "0.15.0"
+futures = { version = "0.3.0", features = ["thread-pool"] }
+jasondb = { version = "0.1.0", path = "../jasondb" }
+native-tls = "0.2.7"
+parking_lot = "0.10.2"
+serde = "1.0"
+serde_json = "1.0"
+tokio = { version = "1.5.0", features = ["full", "tracing"] }
+tokio-util = { version = "0.6.3", features = ["full"] }
+websocket = "0.21.1"
+
+[[bin]]
+name = "jasondb"
+path = "src/main.rs"
