+++ jasondb/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "jasondb"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tar = { version = "0.4.33", default-features = false }
+serde = { version = "1.0", optional = true }
+serde_json = { version = "1.0", optional = true }
+
+[features]
+validation = ["serde", "serde_json"]
+default = ["validation"]
