+++ jasondb/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod database;
+pub mod isam;
+
+#[cfg(test)]
+mod tests;
