+++ jasondb/src/tests/mod.rs
@@ -1,3 +1,2 @@
 pub mod json;
 pub mod main;
-pub mod request;
